+++ hollow/src/main/java/com/netflix/hollow/core/read/engine/set/HollowSetDeltaHistoricalStateCreator.java
@@ -58,7 +58,7 @@ public class HollowSetDeltaHistoricalStateCreator {
     public void populateHistory() {
         populateStats();
 
-        historicalDataElements.setPointerAndSizeArray = new FixedLengthElementArray(historicalDataElements.memoryRecycler, historicalDataElements.bitsPerFixedLengthSetPortion * (historicalDataElements.maxOrdinal + 1));
+        historicalDataElements.setPointerAndSizeArray = new FixedLengthElementArray(historicalDataElements.memoryRecycler, historicalDataElements.bitsPerFixedLengthSetPortion * ((long)historicalDataElements.maxOrdinal + 1));
         historicalDataElements.elementArray = new FixedLengthElementArray(historicalDataElements.memoryRecycler, historicalDataElements.bitsPerElement * historicalDataElements.totalNumberOfBuckets);
 
         iter.reset();
